// src/enums/FOUR_CC.ts
export enum FOUR_CC {
    DDS = 0x20534444, // "DDS ", last character is empty space

    DXT1 = 827611204,
    DXT2 = 844388420,
    DXT3 = 861165636,
    DXT4 = 877942852,
    DXT5 = 894720068,
    ATI1 = 826889281,
    ATI2 = 843666497,
    DX10 = 808540228,

    R8G8_B8G8 = 1195525970, // legacy UYVY-packed
    G8R8_G8B8 = 1111970375, // legacy YUY2-packed
};